chore: move nodetypes to node-api
