+++ crates/node-builder/src/lib.rs
@@ -34,3 +34,6 @@ pub use reth_node_core::cli::config::{
 
 // re-export the core config for convenience
 pub use reth_node_core::node_config::NodeConfig;
+
+// re-export API types for convenience
+pub use reth_node_api::*;
